# backend/openapi.yaml
openapi: 3.0.3
info:
  title: ''
  version: 0.0.0
paths:
...
  /api/onboarding/step/:
    get:
      operationId: onboarding_step_retrieve
      tags:
      - onboarding
      security:
      - jwtAuth: []
      responses:
        '200':
          description: No response body
    post:
      operationId: onboarding_step_create
      tags:
      - onboarding
      security:
      - jwtAuth: []
      responses:
        '200':
          description: No response body
...
components:
...
  securitySchemes:
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT